server/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart
langchain>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
unstructured>=0.10.0
python-magic>=0.4.27
python-docx>=0.8.11
PyPDF2>=3.0.0
pdfplumber>=0.10.2
docx2txt>=0.8
markdown>=3.4.0
python-frontmatter>=1.0.0
beautifulsoup4>=4.12.0 